finish `O(lgn)` fast `fib` function
Closes #1

+++ examples/fib.rs
@@ -3,7 +3,7 @@ use structopt::StructOpt;
 
 #[derive(StructOpt, Debug)]
 struct Opt {
-    input: u64,
+    input: usize,
 }
 
 fn main() {
